bulletproofs: encapsulate inner product proof rather than giving scalars to the caller
This costs a byte of space on most rangeproofs but it makes the code much simpler (and
this byte can be recovered by a post-processing step if it's really so important). It
allows to benchmark the inner product proof directly and also to modify the proof to
have a variable number of scalars, which will be needed for Benedikt's optimization
and to get non-powers-of-2.